Change ShowcaseMap.radius from double to int

+++ lib/widgets/map/ready_button.dart
@@ -12,7 +12,7 @@ class MapReadyButton extends StatefulWidget {
 
   final LatLng center;
   final double zoom;
-  final double radius;
+  final int radius;
   final List<MapSaveMode> saveModes;
 
   @override
